+++ src/containers/AddFunds/CoinSelectorModal/CoinSelectorModal.tsx
@@ -1,23 +1,15 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Text } from '@components';
-import { network } from '@models/network';
 import { useState } from '@hookstate/core';
 import { globalWalletState } from '@stores/WalletStore';
+import { useNativeToken } from '@hooks';
 import CoinCard from './CoinCard/CoinCard';
 
 const CoinSelector: React.FC<{ onSelect: Function }> = ({ onSelect }) => {
 	const {
-		network: { nativeToken: token, topUpToken }
+		network: { topUpToken }
 	} = useState(globalWalletState()).value;
-	const [nativeToken, setNativeToken] = React.useState<{ symbol: string; name: string }>();
-	useEffect(() => {
-		const fetchDefaultToken = async () => {
-			const { nativeToken: native } = await network();
-			setNativeToken(native);
-		};
-
-		fetchDefaultToken();
-	}, [token]);
+	const { nativeToken, name } = useNativeToken();
 
 	return (
 		<>
@@ -30,7 +22,7 @@ const CoinSelector: React.FC<{ onSelect: Function }> = ({ onSelect }) => {
 					coin={{
 						symbol: nativeToken.symbol,
 						image: nativeToken.symbol.toLowerCase(),
-						name: nativeToken.name
+						name: name!
 					}}
 					onSelect={onSelect}
 				/>
