+++ src/screens/DepositScreen/Deposit/Deposit.hooks.ts
@@ -1,18 +1,18 @@
 import React, { useEffect, useCallback } from 'react';
 import { getProvider } from '@models/wallet';
-import { NativeTokens, nativeTokens, ParaswapToken } from '@models/token';
+import { ParaswapToken } from '@models/token';
 import { globalWalletState } from '@stores/WalletStore';
 import { globalDepositState } from '@stores/DepositStore';
 import { globalExchangeState } from '@stores/ExchangeStore';
 import { aaveMarketTokenToParaswapToken, depositTransaction } from '@models/deposit';
-import { useNavigation, useTokens, useAmplitude, useAuthentication } from '@hooks';
-import { network } from '@models/network';
+import { useNavigation, useTokens, useAmplitude, useAuthentication, useNativeToken } from '@hooks';
 import { Wallet } from 'ethers';
 import { useState } from '@hookstate/core';
 import { Keyboard } from 'react-native';
 import Logger from '@utils/logger';
 
 export const useDeposit = () => {
+	const { nativeToken } = useNativeToken();
 	const { track } = useAmplitude();
 	const navigation = useNavigation();
 	const { tokens } = useTokens();
@@ -20,7 +20,6 @@ export const useDeposit = () => {
 	const { market } = globalDepositState().value;
 	const { gas } = useState(globalExchangeState()).value;
 	const { gweiValue = 0 } = gas || {};
-	const [nativeToken, setNativeToken] = React.useState<ParaswapToken>();
 	const [token] = React.useState<ParaswapToken>(aaveMarketTokenToParaswapToken(market));
 	const [tokenBalance, setTokenBalance] = React.useState('0');
 	const [amount, setAmount] = React.useState('0');
@@ -110,18 +109,6 @@ export const useDeposit = () => {
 			}
 		});
 	};
-
-	useEffect(() => {
-		const loadNativeToken = async () => {
-			const {
-				nativeToken: { symbol: nativeTokenSymbol }
-			} = await network();
-			const native = nativeTokens[nativeTokenSymbol as keyof NativeTokens];
-			setNativeToken(native);
-		};
-
-		loadNativeToken();
-	}, []);
 
 	useEffect(() => {
 		if (token && tokens && tokens.length > 0) {
