+++ src/screens/ExchangeScreen/ExchangeScreen.hooks.ts
@@ -1,18 +1,17 @@
 import React, { useEffect, createRef, useCallback } from 'react';
 import { TextInput, Keyboard } from 'react-native';
-import { useTokens, useNavigation } from '@hooks';
+import { useTokens, useNavigation, useNativeToken } from '@hooks';
 import { useState, State } from '@hookstate/core';
 import { BigNumber, utils } from 'ethers';
-import { ParaswapToken, Quote, getExchangePrice, nativeTokens, NativeTokens, ExchangeParams } from '@models/token';
-import { network } from '@models/network';
+import { ParaswapToken, Quote, getExchangePrice, ExchangeParams } from '@models/token';
 import { ExchangeState, Conversion, globalExchangeState } from '@stores/ExchangeStore';
 import Logger from '@utils/logger';
 
 export const useExchangeScreen = () => {
+	const { nativeToken } = useNativeToken();
 	const navigation = useNavigation();
 	const exchange: State<ExchangeState> = useState(globalExchangeState());
 	const [searchVisible, setSearchVisible] = React.useState(false);
-	const [nativeToken, setNativeToken] = React.useState<ParaswapToken>();
 	const [fromToken, setFromToken] = React.useState<ParaswapToken>({} as ParaswapToken);
 	const [toToken, setToToken] = React.useState<ParaswapToken>();
 	const [loadingPrices, setLoadingPrices] = React.useState(false);
@@ -213,20 +212,6 @@ export const useExchangeScreen = () => {
 			updateToToken(backup);
 		}
 	};
-
-	useEffect(() => {
-		const loadNativeToken = async () => {
-			const {
-				nativeToken: { symbol: nativeTokenSymbol }
-			} = await network();
-			const native = nativeTokens[nativeTokenSymbol as keyof NativeTokens];
-			setFromToken(native);
-			setNativeToken(native);
-		};
-
-		exchange.set({} as ExchangeState);
-		loadNativeToken();
-	}, []);
 
 	useEffect(() => {
 		if (walletTokens) {
